Fix import ol-ext & move serch input into to Map component
