+++ pages/index.vue
@@ -1,20 +1,17 @@
 <template>
   <div>
     <Introduction class="container mx-auto px-4" />
-    <Search />
     <Map :events="events" />
   </div>
 </template>
 
 <script>
 import Introduction from '~/components/introduction.vue'
-import Search from '~/components/search.vue'
 import Map from '~/components/map.vue'
 
 export default {
   components: {
     Introduction,
-    Search,
     Map,
   },
